requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
argon2-cffi==23.1.0
pydantic[email]==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0
aiofiles==23.2.1
httpx==0.25.1
tenacity==8.2.3
python-magic==0.4.27
filetype==1.2.0
python-magic-bin==0.4.14; platform_system == "Windows"